feat(mcp): promote MCP Registry out of beta
The beta registry at /mcp/registry/beta becomes the registry at
/mcp/registry for everyone (no longer gated behind ARCHESTRA_BETA):

- move the beta list/new/[id]/catalog routes up to /mcp/registry and
  merge beta/_parts into registry/_parts, replacing the legacy forks
  (InternalMCPCatalog, mcp-server-card, archestra-catalog-tab,
  edit-catalog-dialog, manage-users-dialog — identical exported APIs)
- delete the legacy list page and its dialog-based create flow
  (create-catalog-dialog); the layout's "Add MCP Server" button now
  routes to the /mcp/registry/new setup wizard and the Beta title
  badge is gone
- permanent redirect /mcp/registry/beta/:path* -> /mcp/registry/:path*
  so existing bookmarks and links keep working
- update e2e helpers/specs from the legacy create dialog to the routed
  wizard flow (source step, page-scoped form, install from the Test
  connection step)
