+++ platform/e2e-tests/tests/mcp-install.spec.ts
@@ -4,9 +4,9 @@ import { type Page, test } from "../fixtures";
 import {
   clickButton,
   closeOpenDialogs,
+  goToAddMcpServerPage,
   goToMcpRegistry,
   installMcpServer,
-  openAddMcpServerDialog,
   submitAddServer,
   waitForInstallDialog,
   waitForMcpServerCard,
@@ -36,8 +36,8 @@ test.describe("MCP Install", () => {
 
     await goToMcpRegistry(adminPage);
 
-    // Open "Add MCP Server" dialog
-    await openAddMcpServerDialog(adminPage);
+    // "Add MCP Server" navigates to the routed setup wizard
+    await goToAddMcpServerPage(adminPage);
 
     // Browse online catalog to search for context7
     await adminPage
@@ -58,8 +58,10 @@ test.describe("MCP Install", () => {
     // Submit the pre-filled form to add server to registry
     await submitAddServer(adminPage);
 
-    // Install dialog opens automatically after adding to registry
-    // Wait for the install dialog to be visible
+    // Creating the item continues the setup wizard on its "Test connection"
+    // step — kick off the install from there, which opens the install dialog
+    // to collect the prompted credentials.
+    await clickButton({ page: adminPage, options: { name: "Install" } });
     await waitForInstallDialog(adminPage, { titlePattern: /Install -/ });
 
     // fill the api key (just fake value)
@@ -71,6 +73,7 @@ test.describe("MCP Install", () => {
     await installMcpServer(adminPage);
 
     // Wait for the card to appear in the registry after installation
+    await goToMcpRegistry(adminPage);
     await waitForMcpServerCard(adminPage, CONTEXT7_CATALOG_ITEM_NAME);
 
     // Check that tools are discovered
@@ -149,8 +152,8 @@ rl.on("line", (line) => {
     // ========================================
     // STEP 1: Create MCP server with bogus image
     // ========================================
-    await clickButton({ page: adminPage, options: { name: "Add MCP Server" } });
-    await adminPage.waitForLoadState("domcontentloaded");
+    await goToAddMcpServerPage(adminPage);
+    await adminPage.getByRole("button", { name: "Start from scratch" }).click();
 
     await adminPage
       .getByRole("button", {
@@ -174,15 +177,14 @@ rl.on("line", (line) => {
     await clickButton({ page: adminPage, options: { name: "Add Server" } });
     await adminPage.waitForLoadState("domcontentloaded");
 
-    // Wait for install dialog and install the server
-    await adminPage
-      .getByRole("dialog")
-      .filter({ hasText: /Install -/ })
-      .waitFor({ state: "visible", timeout: 30000 });
+    // Creating the item lands on the setup wizard's "Test connection" step.
+    // No credentials are prompted for this server, so "Install" starts the
+    // install directly (no dialog).
     await clickButton({ page: adminPage, options: { name: "Install" } });
     await adminPage.waitForLoadState("domcontentloaded");
 
-    // Wait for the server card to appear
+    // Wait for the server card to appear in the registry
+    await goToMcpRegistry(adminPage);
     const serverCard = adminPage.getByTestId(
       `${E2eTestId.McpServerCard}-${CATALOG_ITEM_NAME}`,
     );
