+++ platform/frontend/src/app/mcp/registry/[id]/edit/page.client.tsx
@@ -38,7 +38,7 @@ export function McpCatalogItemEditPage({ id }: { id: string }) {
         className="-ml-2 text-muted-foreground"
         asChild
       >
-        <Link href={`/mcp/registry/beta/${id}`}>
+        <Link href={`/mcp/registry/${id}`}>
           <ArrowLeft className="h-4 w-4" />
           Back to server
         </Link>
@@ -179,9 +179,7 @@ function SetupWizard({ item }: { item: CatalogItem }) {
               </Button>
             )
           ) : (
-            <Button
-              onClick={() => router.push(`/mcp/registry/beta/${item.id}`)}
-            >
+            <Button onClick={() => router.push(`/mcp/registry/${item.id}`)}>
               Finish
             </Button>
           )}
