+++ platform/frontend/src/app/mcp/registry/[id]/edit/page.tsx
