+++ platform/frontend/src/app/mcp/registry/[id]/page.tsx
