+++ platform/frontend/src/app/mcp/registry/_parts/InternalMCPCatalog.tsx
@@ -3,14 +3,12 @@
 import {
   ARCHESTRA_MCP_CATALOG_ID,
   isPlaywrightCatalogItem,
-  MCP_CATALOG_INSTALL_QUERY_PARAM,
   MCP_CATALOG_REAUTH_QUERY_PARAM,
   MCP_CATALOG_SERVER_QUERY_PARAM,
 } from "@archestra/shared";
-import { useQueryClient } from "@tanstack/react-query";
 import { Search } from "lucide-react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { toast } from "sonner";
 import {
   LabelFilterBadges,
@@ -28,31 +26,15 @@ import { Button } from "@/components/ui/button";
 import { useHasPermissions, useSession } from "@/lib/auth/auth.query";
 import { useInitiateOAuth } from "@/lib/auth/oauth.query";
 import {
-  clearInstallationCompleteCatalogId,
-  clearPendingAfterEnvVars,
-  getOAuthInstallationCompleteCatalogId,
-  getOAuthPendingAfterEnvVars,
   setOAuthCatalogId,
-  setOAuthEnvironmentValues,
-  setOAuthIsFirstInstallation,
   setOAuthMcpServerId,
-  setOAuthPendingAfterEnvVars,
   setOAuthReturnUrl,
   setOAuthScope,
-  setOAuthServerType,
   setOAuthState,
   setOAuthTeamId,
-  setOAuthUserConfigValues,
 } from "@/lib/auth/oauth-session";
+import { useEnvironments } from "@/lib/environment.query";
 import { useDialogs } from "@/lib/hooks/use-dialog";
-import {
-  clearPendingEnterpriseManagedInstall,
-  type EnterpriseManagedInstallIntent,
-  getPendingEnterpriseManagedInstall,
-  setPendingEnterpriseManagedInstall,
-  useEnterpriseManagedInstallConnectUrl,
-} from "@/lib/mcp/enterprise-managed-install-auth";
-import { useMcpRegistryServer } from "@/lib/mcp/external-mcp-catalog.query";
 import {
   useInternalMcpCatalog,
   useMcpCatalogLabelKeys,
@@ -60,7 +42,6 @@ import {
   useReinstallInternalMcpCatalogItem,
 } from "@/lib/mcp/internal-mcp-catalog.query";
 import {
-  useInstallMcpServer,
   useMcpDeploymentStatuses,
   useMcpInstallationStatusCacheSync,
   useMcpServers,
@@ -68,35 +49,37 @@ import {
   useReinstallMcpServer,
 } from "@/lib/mcp/mcp-server.query";
 import { buildRemoteInstallCredentialPayload } from "@/lib/mcp/remote-install-payload";
-import websocketService from "@/lib/websocket/websocket";
-import { CreateCatalogDialog } from "./create-catalog-dialog";
+import { useDefaultEnvironment } from "@/lib/organization.query";
+import { resolveCatalogEnvironmentLabel } from "./catalog-environment-label";
 import { CustomServerRequestDialog } from "./custom-server-request-dialog";
-import { DeleteCatalogDialog } from "./delete-catalog-dialog";
-import { DetailsDialog } from "./details-dialog";
-import { EditCatalogDialog } from "./edit-catalog-dialog";
 import {
   LocalServerInstallDialog,
   type LocalServerInstallResult,
 } from "./local-server-install-dialog";
 import { ManageUsersDialog } from "./manage-users-dialog";
-import type { McpCatalogFormValues } from "./mcp-catalog-form.types";
-import { buildCloneFormValues } from "./mcp-catalog-form.utils";
 import {
   type CatalogItem,
-  cardVariantForServerType,
   type InstalledServer,
   McpServerCard,
 } from "./mcp-server-card";
 import {
-  NoAuthInstallDialog,
-  type NoAuthInstallResult,
-} from "./no-auth-install-dialog";
+  emptyRegistryFilters,
+  type FilterGroup,
+  type FilterOption,
+  RegistryFilterChips,
+  RegistryFilterDropdown,
+  type RegistryFilters,
+  RegistrySortMenu,
+  type SortKey,
+  STATUS_OPTIONS,
+} from "./registry-list-controls";
 import { ReinstallConfirmationDialog } from "./reinstall-confirmation-dialog";
 import {
   RemoteServerInstallDialog,
   type RemoteServerInstallResult,
 } from "./remote-server-install-dialog";
 import type { McpServerInstallScope } from "./select-mcp-server-credential-type-and-teams";
+import { useCatalogInstall } from "./use-catalog-install";
 
 export function InternalMCPCatalog({
   initialData,
@@ -112,22 +95,25 @@ export function InternalMCPCatalog({
   // Get search query from URL
   const searchQueryFromUrl = searchParams.get("search") || "";
 
-  const { data: catalogItems } = useInternalMcpCatalog({ initialData });
-  const [installingServerIds, setInstallingServerIds] = useState<Set<string>>(
-    new Set(),
-  );
-  const [restartingServerIds, setRestartingServerIds] = useState<Set<string>>(
-    new Set(),
-  );
-  // Track server IDs that are first-time installations (for auto-opening assignments dialog)
-  const [firstInstallationServerIds, setFirstInstallationServerIds] = useState<
-    Set<string>
-  >(new Set());
+  const { data: catalogItems } = useInternalMcpCatalog({
+    initialData,
+  });
   const { data: installedServers } = useMcpServers({
     initialData: initialInstalledServers,
   });
   useMcpInstallationStatusCacheSync();
-  const installMutation = useInstallMcpServer();
+
+  // Shared install flow (install / add-personal / add-shared / add-org,
+  // remote / local / no-auth, OAuth, enterprise guard). Reinstall and reauth
+  // live below and reuse the polling set + installingItemId exposed here.
+  const install = useCatalogInstall();
+  const {
+    installingItemId,
+    installingServerIds,
+    setInstallingServerIds,
+    setInstallingItemId,
+  } = install;
+
   const reinstallMutation = useReinstallMcpServer();
   // When the card requests an admin combined reinstall, remember which
   // catalog id needs its shared pod recreated *after* the per-install
@@ -141,38 +127,42 @@ export function InternalMCPCatalog({
   const deploymentStatuses = useMcpDeploymentStatuses();
   const { data: session } = useSession();
   const currentUserId = session?.user?.id;
+  const { data: environmentList } = useEnvironments();
+  const defaultEnvironment = useDefaultEnvironment();
+
+  const [sort, setSort] = useState<SortKey>("name-asc");
+  const [filters, setFilters] = useState<RegistryFilters>(emptyRegistryFilters);
+  const toggleFilter = useCallback((group: FilterGroup, value: string) => {
+    setFilters((prev) => {
+      const next = new Set(prev[group]);
+      if (next.has(value)) next.delete(value);
+      else next.add(value);
+      return { ...prev, [group]: next };
+    });
+  }, []);
+  const removeFilter = useCallback((group: FilterGroup, value: string) => {
+    setFilters((prev) => {
+      const next = new Set(prev[group]);
+      next.delete(value);
+      return { ...prev, [group]: next };
+    });
+  }, []);
+  const clearAdvancedFilters = useCallback(
+    () => setFilters(emptyRegistryFilters()),
+    [],
+  );
 
   const { isDialogOpened, openDialog, closeDialog } = useDialogs<
-    | "create"
     | "custom-request"
-    | "edit"
-    | "delete"
     | "remote-install"
     | "local-install"
     | "oauth"
-    | "no-auth"
     | "reinstall"
     | "manage"
   >();
 
-  const [editingItem, setEditingItem] = useState<CatalogItem | null>(null);
-  const [cloneValues, setCloneValues] = useState<McpCatalogFormValues | null>(
-    null,
-  );
-  const [cloneSourceId, setCloneSourceId] = useState<string | null>(null);
-  const [deletingItem, setDeletingItem] = useState<CatalogItem | null>(null);
-  const [installingItemId, setInstallingItemId] = useState<string | null>(null);
-
   // Deep-link manage connections dialog state
   const [manageCatalogId, setManageCatalogId] = useState<string | null>(null);
-  // Pre-selected team ID when adding a shared connection from manage dialog
-  const [preselectedTeamId, setPreselectedTeamId] = useState<string | null>(
-    null,
-  );
-  // When true, install dialog hides the team selector (personal connection only)
-  const [installPersonalOnly, setInstallPersonalOnly] = useState(false);
-  // When true, install dialog forces the organization-wide scope
-  const [installOrgOnly, setInstallOrgOnly] = useState(false);
 
   // Update URL when search query changes (debounced via DebouncedInput)
   const handleSearchChange = useCallback(
@@ -197,8 +187,6 @@ export function InternalMCPCatalog({
   const [reinstallFlaggedTargets, setReinstallFlaggedTargets] = useState<
     Array<{ id: string; name: string }>
   >([]);
-  const [noAuthCatalogItem, setNoAuthCatalogItem] =
-    useState<CatalogItem | null>(null);
   const [localServerCatalogItem, setLocalServerCatalogItem] =
     useState<CatalogItem | null>(null);
   // Track server ID when reinstalling (vs new installation)
@@ -215,195 +203,18 @@ export function InternalMCPCatalog({
   >(undefined);
   // Track server ID for re-authentication (preserves tool assignments)
   const [reauthServerId, setReauthServerId] = useState<string | null>(null);
-  const [detailsServerName, setDetailsServerName] = useState<string | null>(
-    null,
-  );
-
-  // The connection being re-authenticated keeps its existing scope, so the
-  // install dialogs need its scope/team to lock the selector (otherwise the
-  // selector treats re-auth as a fresh install and disables the already-used
-  // scope, leaving the owner unable to re-authenticate their own connection).
-  const reauthServer = useMemo(
-    () =>
-      reauthServerId
-        ? installedServers?.find((server) => server.id === reauthServerId)
-        : undefined,
-    [reauthServerId, installedServers],
-  );
-  const reauthExistingScope: McpServerInstallScope | undefined = reauthServerId
-    ? (reauthServer?.scope ?? (reauthServer?.teamId ? "team" : "personal"))
-    : undefined;
-  const reauthExistingTeamId: string | null | undefined = reauthServerId
-    ? (reauthServer?.teamId ?? null)
-    : undefined;
-  const { data: detailsServerData } = useMcpRegistryServer(detailsServerName);
 
   const { data: _userIsMcpServerAdmin } = useHasPermissions({
     mcpServerInstallation: ["admin"],
   });
 
-  const queryClient = useQueryClient();
-  const getEnterpriseManagedInstallConnectUrl =
-    useEnterpriseManagedInstallConnectUrl();
-
-  const ensureEnterpriseManagedInstallAuth = useCallback(
-    async (
-      catalogItem: CatalogItem,
-      intent: EnterpriseManagedInstallIntent,
-    ): Promise<boolean> => {
-      const connectUrl = await getEnterpriseManagedInstallConnectUrl({
-        catalogItem,
-        redirectTo: "/mcp/registry",
-      });
-      if (!connectUrl) {
-        return true;
-      }
-
-      setPendingEnterpriseManagedInstall(intent);
-      window.location.assign(connectUrl);
-      return false;
-    },
-    [getEnterpriseManagedInstallConnectUrl],
-  );
-
-  // Remove servers from installing set when installation completes (success or error)
+  // Deep-link: auto-open install dialog when ?install={catalogId} is present.
+  // Optional &scope=personal|team|org (and &team={teamId} for team scope)
+  // pre-target the connection — used by the item detail page's add-connection
+  // actions. Owned by the shared install hook.
+  // biome-ignore lint/correctness/useExhaustiveDependencies: only trigger on searchParams/catalogItems changes, installFromSearchParams is a stable callback
   useEffect(() => {
-    if (installedServers && installingServerIds.size > 0) {
-      const completedServerIds = Array.from(installingServerIds).filter(
-        (serverId) => {
-          const server = installedServers.find((s) => s.id === serverId);
-          return (
-            server &&
-            (server.localInstallationStatus === "success" ||
-              server.localInstallationStatus === "error")
-          );
-        },
-      );
-
-      if (completedServerIds.length > 0) {
-        setInstallingServerIds((prev) => {
-          const newSet = new Set(prev);
-          for (const id of completedServerIds) {
-            newSet.delete(id);
-          }
-          return newSet;
-        });
-
-        // Show toasts for completed installations and invalidate tools queries
-        completedServerIds.forEach((serverId) => {
-          const server = installedServers.find((s) => s.id === serverId);
-          if (server) {
-            if (server.localInstallationStatus === "success") {
-              if (!restartingServerIds.has(serverId)) {
-                toast.success(`Successfully installed ${server.name}`);
-              }
-              // Force immediate deployment status refresh via WebSocket
-              websocketService.send({
-                type: "subscribe_mcp_deployment_statuses",
-                payload: {},
-              });
-              // Invalidate tools queries to update "Tools assigned" count
-              queryClient.invalidateQueries({
-                queryKey: ["mcp-servers", server.id, "tools"],
-              });
-              queryClient.invalidateQueries({ queryKey: ["tools"] });
-              queryClient.invalidateQueries({
-                queryKey: ["tools", "unassigned"],
-              });
-              // Invalidate catalog tools so the manage-tools dialog shows discovered tools
-              if (server.catalogId) {
-                queryClient.invalidateQueries({
-                  queryKey: ["mcp-catalog", server.catalogId, "tools"],
-                });
-
-                // Remove from first installation tracking
-                if (firstInstallationServerIds.has(serverId)) {
-                  setFirstInstallationServerIds((prev) => {
-                    const newSet = new Set(prev);
-                    newSet.delete(serverId);
-                    return newSet;
-                  });
-                }
-              }
-            }
-            if (
-              restartingServerIds.has(serverId) &&
-              (server.localInstallationStatus === "success" ||
-                server.localInstallationStatus === "error")
-            ) {
-              setRestartingServerIds((prev) => {
-                const newSet = new Set(prev);
-                newSet.delete(serverId);
-                return newSet;
-              });
-            }
-            // Note: No error toast - the error banner on the card provides feedback
-          }
-        });
-      }
-    }
-  }, [
-    installedServers,
-    installingServerIds,
-    restartingServerIds,
-    queryClient,
-    firstInstallationServerIds,
-  ]);
-
-  // Resume polling for pending installations after page refresh
-  useEffect(() => {
-    if (installedServers) {
-      const pendingServers = installedServers.filter(
-        (s) =>
-          s.localInstallationStatus === "pending" ||
-          s.localInstallationStatus === "discovering-tools",
-      );
-      if (pendingServers.length > 0) {
-        setInstallingServerIds(new Set(pendingServers.map((s) => s.id)));
-      }
-    }
-  }, [installedServers]);
-
-  // Listen for create event from layout header button
-  useEffect(() => {
-    const handler = () => openDialog("create");
-    window.addEventListener("mcp-registry:create", handler);
-    return () => window.removeEventListener("mcp-registry:create", handler);
-  }, [openDialog]);
-
-  // Clear OAuth installation completion state
-  useEffect(() => {
-    const oauthCatalogId = getOAuthInstallationCompleteCatalogId();
-    if (oauthCatalogId) {
-      clearInstallationCompleteCatalogId();
-    }
-  }, []);
-
-  // Deep-link: auto-open install dialog when ?install={catalogId} is present
-  // biome-ignore lint/correctness/useExhaustiveDependencies: only trigger on searchParams/catalogItems changes, other deps are stable callbacks
-  useEffect(() => {
-    const installCatalogId = searchParams.get(MCP_CATALOG_INSTALL_QUERY_PARAM);
-    if (!installCatalogId || !catalogItems) return;
-
-    const catalogItem = catalogItems.find(
-      (item) => item.id === installCatalogId,
-    );
-    if (!catalogItem) return;
-
-    // Clear the install param from URL to prevent re-triggering on refresh
-    const params = new URLSearchParams(searchParams.toString());
-    params.delete(MCP_CATALOG_INSTALL_QUERY_PARAM);
-    const newUrl = params.toString()
-      ? `${pathname}?${params.toString()}`
-      : pathname;
-    router.replace(newUrl, { scroll: false });
-
-    // Trigger the appropriate install dialog
-    if (catalogItem.serverType === "local") {
-      handleInstallLocalServer(catalogItem);
-    } else {
-      handleInstallRemoteServer(catalogItem, false);
-    }
+    install.installFromSearchParams();
   }, [searchParams, catalogItems]);
 
   // Deep-link: handle ?reauth={catalogId} with optional ?server={serverId}
@@ -484,395 +295,42 @@ export function InternalMCPCatalog({
     }
   };
 
-  const handleInstallRemoteServer = async (
-    catalogItem: CatalogItem,
-    _teamMode: boolean,
-    options?: {
-      preserveInstallTarget?: boolean;
-      scope?: McpServerInstallScope;
-      teamId?: string;
-    },
-  ) => {
-    if (!options?.preserveInstallTarget) {
-      setPreselectedTeamId(null);
-      setInstallPersonalOnly(false);
-      setInstallOrgOnly(false);
-    }
+  // OAuth confirm for re-authentication (install OAuth lives in useCatalogInstall).
+  const handleReauthOAuthConfirm = async (result: OAuthInstallResult) => {
+    if (!selectedCatalogItem) return;
 
-    const scope = options?.scope
-      ? options.scope
-      : installOrgOnly
-        ? "org"
-        : preselectedTeamId
-          ? "team"
-          : installPersonalOnly
-            ? "personal"
-            : undefined;
-    const teamId = options?.teamId ?? preselectedTeamId ?? undefined;
-    if (
-      !(await ensureEnterpriseManagedInstallAuth(catalogItem, {
-        action: "open-remote",
-        catalogId: catalogItem.id,
-        scope,
-        ...(teamId ? { teamId } : {}),
-      }))
-    ) {
-      return;
-    }
+    try {
+      const { authorizationUrl, state } =
+        await initiateOAuthMutation.mutateAsync({
+          catalogId: selectedCatalogItem.id,
+        });
 
-    const hasUserConfig =
-      catalogItem.userConfig && Object.keys(catalogItem.userConfig).length > 0;
+      setOAuthState(state);
+      setOAuthCatalogId(selectedCatalogItem.id);
+      setOAuthTeamId(result.scope === "team" ? (result.teamId ?? null) : null);
+      setOAuthScope(result.scope);
 
-    // Check if this server requires OAuth authentication if there is no user config
-    if (!hasUserConfig && catalogItem.oauthConfig) {
-      setSelectedCatalogItem(catalogItem);
-      openDialog("oauth");
-      return;
-    }
-
-    setSelectedCatalogItem(catalogItem);
-    openDialog("remote-install");
-  };
-
-  const handleInstallLocalServer = async (
-    catalogItem: CatalogItem,
-    options?: {
-      preserveInstallTarget?: boolean;
-      scope?: McpServerInstallScope;
-      teamId?: string;
-    },
-  ) => {
-    if (!options?.preserveInstallTarget) {
-      setPreselectedTeamId(null);
-      setInstallPersonalOnly(false);
-      setInstallOrgOnly(false);
-    }
-
-    const scope = options?.scope
-      ? options.scope
-      : installOrgOnly
-        ? "org"
-        : preselectedTeamId
-          ? "team"
-          : installPersonalOnly
-            ? "personal"
-            : undefined;
-    const teamId = options?.teamId ?? preselectedTeamId ?? undefined;
-    if (
-      !(await ensureEnterpriseManagedInstallAuth(catalogItem, {
-        action: "open-local",
-        catalogId: catalogItem.id,
-        scope,
-        ...(teamId ? { teamId } : {}),
-      }))
-    ) {
-      return;
-    }
-
-    // Check if this local server requires OAuth authentication
-    if (catalogItem.oauthConfig) {
-      // Check if there are prompted env vars that need collecting first
-      const promptedEnvVars =
-        catalogItem.localConfig?.environment?.filter(
-          (env) => env.promptOnInstallation === true,
-        ) || [];
-
-      const promptableUserConfig = Object.values(
-        catalogItem.userConfig ?? {},
-      ).filter((field) => field.promptOnInstallation !== false);
-
-      if (promptedEnvVars.length > 0 || promptableUserConfig.length > 0) {
-        // Has prompted env vars or promptable user-config - open local install dialog first to collect them,
-        // then initiate OAuth after dialog confirm
-        setLocalServerCatalogItem(catalogItem);
-        setOAuthPendingAfterEnvVars(true);
-        openDialog("local-install");
-      } else {
-        // No env vars needed - go straight to OAuth flow
-        // Store server type so OAuth callback knows this is a local server
-        setOAuthServerType("local");
-        setSelectedCatalogItem(catalogItem);
-        openDialog("oauth");
+      if (reauthServerId) {
+        setOAuthMcpServerId(reauthServerId);
+        setOAuthReturnUrl(window.location.href);
+        setReauthServerId(null);
       }
-      return;
-    }
 
-    setLocalServerCatalogItem(catalogItem);
-    openDialog("local-install");
-  };
-
-  const handleInstallPlaywright = async (catalogItem: CatalogItem) => {
-    setInstallingItemId(catalogItem.id);
-    const result = await installMutation.mutateAsync({
-      name: catalogItem.name,
-      catalogId: catalogItem.id,
-      dontShowToast: true,
-    });
-
-    const installedServerId = result?.installedServer?.id;
-    if (installedServerId) {
-      setInstallingServerIds((prev) => new Set(prev).add(installedServerId));
-      const isFirstInstallation = !installedServers?.some(
-        (s) => s.catalogId === catalogItem.id,
+      window.location.href = authorizationUrl;
+    } catch (error) {
+      toast.error(
+        error instanceof Error
+          ? error.message
+          : "Failed to initiate OAuth flow",
       );
-      if (isFirstInstallation) {
-        setFirstInstallationServerIds((prev) =>
-          new Set(prev).add(installedServerId),
-        );
-      }
-    }
-    setInstallingItemId(null);
-  };
-
-  // Check if a catalog item needs any config dialogs, or can be installed directly
-  const canDirectInstall = (catalogItem: CatalogItem) => {
-    if (catalogItem.oauthConfig) return false;
-    if (catalogItem.serverType === "remote") {
-      const hasUserConfig =
-        catalogItem.userConfig &&
-        Object.keys(catalogItem.userConfig).length > 0;
-      return !hasUserConfig;
-    }
-    // Local server: check for prompted env vars or promptable user-config
-    const promptedEnvVars =
-      catalogItem.localConfig?.environment?.filter(
-        (env) => env.promptOnInstallation === true,
-      ) || [];
-    const promptableUserConfig = Object.values(
-      catalogItem.userConfig ?? {},
-    ).filter((field) => field.promptOnInstallation !== false);
-    return promptedEnvVars.length === 0 && promptableUserConfig.length === 0;
-  };
-
-  // Install directly without opening a dialog (works for personal, team, and org)
-  const handleDirectInstall = async (
-    catalogItem: CatalogItem,
-    target?: {
-      teamId?: string;
-      scope?: McpServerInstallScope;
-    },
-  ) => {
-    const scope: McpServerInstallScope =
-      target?.scope ?? (target?.teamId ? "team" : "personal");
-    if (
-      !(await ensureEnterpriseManagedInstallAuth(catalogItem, {
-        action: "direct",
-        catalogId: catalogItem.id,
-        scope,
-        ...(scope === "team" && target?.teamId
-          ? { teamId: target.teamId }
-          : {}),
-      }))
-    ) {
-      return;
-    }
-
-    setInstallingItemId(catalogItem.id);
-    const result = await installMutation.mutateAsync({
-      name: catalogItem.name,
-      catalogId: catalogItem.id,
-      scope,
-      ...(scope === "team" && target?.teamId ? { teamId: target.teamId } : {}),
-      dontShowToast: true,
-    });
-
-    const installedServerId = result?.installedServer?.id;
-    if (installedServerId) {
-      setInstallingServerIds((prev) => new Set(prev).add(installedServerId));
-      const isFirstInstallation = !installedServers?.some(
-        (s) => s.catalogId === catalogItem.id,
-      );
-      if (isFirstInstallation) {
-        setFirstInstallationServerIds((prev) =>
-          new Set(prev).add(installedServerId),
-        );
-      }
-    }
-    setInstallingItemId(null);
-  };
-
-  // Add personal connection: skip dialog if no config needed, otherwise open dialog with personalOnly
-  const handleAddPersonalConnection = (catalogItem: CatalogItem) => {
-    if (canDirectInstall(catalogItem)) {
-      handleDirectInstall(catalogItem);
-    } else {
-      setInstallPersonalOnly(true);
-      if (catalogItem.serverType === "local") {
-        handleInstallLocalServer(catalogItem, {
-          preserveInstallTarget: true,
-          scope: "personal",
-        });
-      } else {
-        handleInstallRemoteServer(catalogItem, false, {
-          preserveInstallTarget: true,
-          scope: "personal",
-        });
-      }
     }
   };
 
-  // Add shared connection: skip dialog if no config needed, otherwise open dialog with preselected team
-  const handleAddSharedConnection = (
-    catalogItem: CatalogItem,
-    teamId: string,
-  ) => {
-    if (canDirectInstall(catalogItem)) {
-      handleDirectInstall(catalogItem, {
-        teamId,
-        scope: "team",
-      });
-    } else {
-      setPreselectedTeamId(teamId);
-      if (catalogItem.serverType === "local") {
-        handleInstallLocalServer(catalogItem, {
-          preserveInstallTarget: true,
-          scope: "team",
-          teamId,
-        });
-      } else {
-        handleInstallRemoteServer(catalogItem, false, {
-          preserveInstallTarget: true,
-          scope: "team",
-          teamId,
-        });
-      }
-    }
-  };
-
-  // Add organization connection: skip dialog if no config needed, otherwise
-  // open dialog with scope locked to org.
-  const handleAddOrgConnection = (catalogItem: CatalogItem) => {
-    if (canDirectInstall(catalogItem)) {
-      handleDirectInstall(catalogItem, { scope: "org" });
-    } else {
-      setInstallOrgOnly(true);
-      if (catalogItem.serverType === "local") {
-        handleInstallLocalServer(catalogItem, {
-          preserveInstallTarget: true,
-          scope: "org",
-        });
-      } else {
-        handleInstallRemoteServer(catalogItem, false, {
-          preserveInstallTarget: true,
-          scope: "org",
-        });
-      }
-    }
-  };
-
-  // Resume an enterprise-managed install after linking the configured IdP.
-  // biome-ignore lint/correctness/useExhaustiveDependencies: consume the one-shot sessionStorage intent when catalog data becomes available
-  useEffect(() => {
-    if (!catalogItems) return;
-
-    const intent = getPendingEnterpriseManagedInstall();
-    if (!intent) return;
-
-    const catalogItem = catalogItems.find(
-      (item) => item.id === intent.catalogId,
-    );
-    if (!catalogItem) return;
-
-    clearPendingEnterpriseManagedInstall();
-
-    setPreselectedTeamId(
-      intent.scope === "team" ? (intent.teamId ?? null) : null,
-    );
-    setInstallPersonalOnly(intent.scope === "personal");
-    setInstallOrgOnly(intent.scope === "org");
-
-    switch (intent.action) {
-      case "direct":
-        void handleDirectInstall(catalogItem, {
-          scope: intent.scope,
-          teamId: intent.teamId,
-        });
-        return;
-      case "open-local":
-        void handleInstallLocalServer(catalogItem, {
-          preserveInstallTarget: true,
-          scope: intent.scope,
-          teamId: intent.teamId,
-        });
-        return;
-      case "open-remote":
-        void handleInstallRemoteServer(catalogItem, false, {
-          preserveInstallTarget: true,
-          scope: intent.scope,
-          teamId: intent.teamId,
-        });
-        return;
-    }
-  }, [catalogItems]);
-
-  const handleNoAuthConfirm = async (result: NoAuthInstallResult) => {
-    if (!noAuthCatalogItem) return;
-
-    const catalogItem = noAuthCatalogItem;
-    setInstallingItemId(catalogItem.id);
-    await installMutation.mutateAsync({
-      name: catalogItem.name,
-      catalogId: result.catalogId,
-      scope: result.scope,
-      teamId:
-        result.scope === "team" ? (result.teamId ?? undefined) : undefined,
-    });
-    closeDialog("no-auth");
-    setNoAuthCatalogItem(null);
-    setInstallingItemId(null);
-  };
-
-  const handleLocalServerInstallConfirm = async (
+  // Re-authentication confirm for local servers (reuses the local install dialog).
+  const handleLocalServerReauthOrReinstallConfirm = async (
     installResult: LocalServerInstallResult,
   ) => {
     if (!localServerCatalogItem) return;
-
-    // Check if OAuth is pending after env vars collection
-    if (getOAuthPendingAfterEnvVars() && localServerCatalogItem.oauthConfig) {
-      clearPendingAfterEnvVars();
-      // Store env vars and server type for use after OAuth callback
-      setOAuthServerType("local");
-      if (
-        installResult.environmentValues &&
-        Object.keys(installResult.environmentValues).length > 0
-      ) {
-        // Security: filter out secret-type env vars from sessionStorage.
-        // In BYOS mode values are vault references (safe). In non-BYOS mode
-        // actual secret values are excluded — they are handled server-side
-        // via secretId or re-prompted on install.
-        const secretKeys = new Set(
-          (localServerCatalogItem.localConfig?.environment ?? [])
-            .filter((e) => e.type === "secret")
-            .map((e) => e.key),
-        );
-        const safeValues = installResult.isByosVault
-          ? installResult.environmentValues
-          : Object.fromEntries(
-              Object.entries(installResult.environmentValues).filter(
-                ([key]) => !secretKeys.has(key),
-              ),
-            );
-        if (Object.keys(safeValues).length > 0) {
-          setOAuthEnvironmentValues(safeValues);
-        }
-      }
-      if (
-        installResult.userConfigValues &&
-        Object.keys(installResult.userConfigValues).length > 0
-      ) {
-        setOAuthUserConfigValues({
-          values: installResult.userConfigValues,
-          userConfig: localServerCatalogItem.userConfig,
-          isByosVault: installResult.isByosVault,
-        });
-      }
-      closeDialog("local-install");
-      // Now initiate OAuth flow
-      setSelectedCatalogItem(localServerCatalogItem);
-      setLocalServerCatalogItem(null);
-      openDialog("oauth");
-      return;
-    }
 
     // Re-authentication mode: update existing server credentials in-place
     if (reauthServerId) {
@@ -890,12 +348,11 @@ export function InternalMCPCatalog({
       return;
     }
 
-    // Check if this is a reinstall (updating existing server) vs new installation
+    // Reinstall mode - apply the submitted values to every flagged install
+    // in the preset family (or just the single one if the card didn't pass a
+    // list). Same env/userConfig bag is applied to each — operators can edit
+    // per-install secrets afterwards from Manage credentials.
     if (reinstallServerId) {
-      // Reinstall mode - apply the submitted values to every flagged install
-      // in the preset family (or just the single one if the card didn't pass a
-      // list). Same env/userConfig bag is applied to each — operators can edit
-      // per-install secrets afterwards from Manage credentials.
       const targetIds =
         reinstallFlaggedTargets.length > 0
           ? reinstallFlaggedTargets.map((t) => t.id)
@@ -948,49 +405,10 @@ export function InternalMCPCatalog({
         setReinstallFlaggedTargets([]);
         setPendingCatalogReinstallId(null);
       }
-      return;
     }
-
-    // New installation flow
-    // Check if this is the first installation for this catalog item
-    const isFirstInstallation = !installedServers?.some(
-      (s) => s.catalogId === localServerCatalogItem.id,
-    );
-
-    setInstallingItemId(localServerCatalogItem.id);
-    const result = await installMutation.mutateAsync({
-      name: localServerCatalogItem.name,
-      catalogId: installResult.catalogId,
-      environmentValues: installResult.environmentValues,
-      userConfigValues: installResult.userConfigValues,
-      isByosVault: installResult.isByosVault,
-      scope: installResult.scope,
-      teamId:
-        installResult.scope === "team"
-          ? (installResult.teamId ?? undefined)
-          : undefined,
-      serviceAccount: installResult.serviceAccount,
-      dontShowToast: true,
-    });
-
-    // Track the installed server for polling
-    const installedServerId = result?.installedServer?.id;
-    if (installedServerId) {
-      setInstallingServerIds((prev) => new Set(prev).add(installedServerId));
-      // Track if this is first installation for opening assignments dialog later
-      if (isFirstInstallation) {
-        setFirstInstallationServerIds((prev) =>
-          new Set(prev).add(installedServerId),
-        );
-      }
-    }
-
-    closeDialog("local-install");
-    setLocalServerCatalogItem(null);
-    setInstallingItemId(null);
   };
 
-  const handleRemoteServerInstallConfirm = async (
+  const handleRemoteServerReauthOrReinstallConfirm = async (
     catalogItem: CatalogItem,
     result: RemoteServerInstallResult,
   ) => {
@@ -1037,59 +455,6 @@ export function InternalMCPCatalog({
           return next;
         });
       }
-      return;
-    }
-
-    setInstallingItemId(catalogItem.id);
-
-    await installMutation.mutateAsync({
-      name: catalogItem.name,
-      catalogId: result.catalogId,
-      ...credentialPayload,
-      scope: result.scope,
-      teamId:
-        result.scope === "team" ? (result.teamId ?? undefined) : undefined,
-    });
-    setInstallingItemId(null);
-  };
-
-  const handleOAuthConfirm = async (result: OAuthInstallResult) => {
-    if (!selectedCatalogItem) return;
-
-    try {
-      // Call backend to initiate OAuth flow
-      const { authorizationUrl, state } =
-        await initiateOAuthMutation.mutateAsync({
-          catalogId: selectedCatalogItem.id,
-        });
-
-      // Store state in session storage for the callback
-      setOAuthState(state);
-      setOAuthCatalogId(selectedCatalogItem.id);
-      setOAuthTeamId(result.scope === "team" ? (result.teamId ?? null) : null);
-      setOAuthScope(result.scope);
-
-      // If re-authenticating via OAuth, store reauth context
-      if (reauthServerId) {
-        setOAuthMcpServerId(reauthServerId);
-        setOAuthReturnUrl(window.location.href);
-        setReauthServerId(null);
-      } else {
-        // Store if this is a first installation (for auto-opening assignments dialog)
-        const isFirstInstallation = !installedServers?.some(
-          (s) => s.catalogId === selectedCatalogItem.id,
-        );
-        setOAuthIsFirstInstallation(isFirstInstallation);
-      }
-
-      // Redirect to OAuth provider
-      window.location.href = authorizationUrl;
-    } catch (error) {
-      toast.error(
-        error instanceof Error
-          ? error.message
-          : "Failed to initiate OAuth flow",
-      );
     }
   };
 
@@ -1318,77 +683,6 @@ export function InternalMCPCatalog({
     }
   };
 
-  const handleClone = (item: CatalogItem) => {
-    setCloneValues(buildCloneFormValues(item));
-    setCloneSourceId(item.id);
-    openDialog("create");
-  };
-
-  const handleCancelInstallation = (serverId: string) => {
-    // Remove server from installing set to stop polling
-    setInstallingServerIds((prev) => {
-      const newSet = new Set(prev);
-      newSet.delete(serverId);
-      return newSet;
-    });
-  };
-
-  const handleRestartPodsStarted = (serverIds: string[]) => {
-    if (serverIds.length === 0) return;
-    setRestartingServerIds((prev) => {
-      const next = new Set(prev);
-      for (const serverId of serverIds) {
-        next.add(serverId);
-      }
-      return next;
-    });
-  };
-
-  const handleRestartPodsFailed = (serverIds: string[]) => {
-    if (serverIds.length === 0) return;
-    setRestartingServerIds((prev) => {
-      const next = new Set(prev);
-      for (const serverId of serverIds) {
-        next.delete(serverId);
-      }
-      return next;
-    });
-  };
-
-  // Capture connected catalog IDs on first load to keep sort order stable.
-  // Only update when the set of catalog IDs changes (new item added/removed),
-  // not when connection status changes (which would cause items to jump around).
-  const connectedCatalogIdsRef = useRef<Set<string> | null>(null);
-  if (connectedCatalogIdsRef.current === null && installedServers) {
-    connectedCatalogIdsRef.current = new Set(
-      installedServers.map((s) => s.catalogId).filter(Boolean) as string[],
-    );
-  }
-
-  const sortInstalledFirst = (items: CatalogItem[]) => {
-    const connectedIds = connectedCatalogIdsRef.current;
-    return [...items].sort((a, b) => {
-      // Primary sort: connected (has installations) first — using stable snapshot
-      const aConnected = connectedIds?.has(a.id) ? 0 : 1;
-      const bConnected = connectedIds?.has(b.id) ? 0 : 1;
-      if (aConnected !== bConnected) return aConnected - bConnected;
-
-      // Secondary sort priority: builtin > remote > local
-      const getPriority = (item: CatalogItem) => {
-        if (item.serverType === "builtin" || isPlaywrightCatalogItem(item.id))
-          return 0;
-        if (item.serverType === "remote") return 1;
-        return 2; // local
-      };
-
-      const priorityDiff = getPriority(a) - getPriority(b);
-      if (priorityDiff !== 0) return priorityDiff;
-
-      // Tertiary sort by createdAt (newest first)
-      return new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime();
-    });
-  };
-
   const filterCatalogItems = (items: CatalogItem[], query: string) => {
     const normalizedQuery = query.trim().toLowerCase();
     if (!normalizedQuery) return items;
@@ -1416,12 +710,100 @@ export function InternalMCPCatalog({
     );
   };
 
-  const allFilteredItems = sortInstalledFirst(
+  // Live connection status (vs the stable snapshot used for the default sort).
+  const connectedCatalogIds = useMemo(
+    () =>
+      new Set(
+        (installedServers ?? [])
+          .map((s) => s.catalogId)
+          .filter(Boolean) as string[],
+      ),
+    [installedServers],
+  );
+  const envLabelByCatalog = useMemo(() => {
+    const envs = environmentList?.environments ?? [];
+    const map = new Map<string, string | null>();
+    for (const it of catalogItems ?? []) {
+      map.set(
+        it.id,
+        it.serverType === "builtin"
+          ? null
+          : (resolveCatalogEnvironmentLabel({
+              environmentId: it.environmentId,
+              environments: envs,
+              defaultEnvironmentName: defaultEnvironment.name,
+            }) ?? defaultEnvironment.name),
+      );
+    }
+    return map;
+  }, [catalogItems, environmentList, defaultEnvironment.name]);
+
+  const environmentOptions: FilterOption[] = useMemo(() => {
+    const set = new Set<string>();
+    envLabelByCatalog.forEach((label) => {
+      if (label) set.add(label);
+    });
+    return [...set].sort().map((value) => ({ value, label: value }));
+  }, [envLabelByCatalog]);
+  const authorOptions: FilterOption[] = useMemo(() => {
+    const set = new Set<string>();
+    for (const it of catalogItems ?? []) {
+      if (it.authorName) set.add(it.authorName);
+    }
+    return [...set].sort().map((value) => ({ value, label: value }));
+  }, [catalogItems]);
+  const matchesAdvancedFilters = (item: CatalogItem) => {
+    if (filters.status.size > 0) {
+      const installed = connectedCatalogIds.has(item.id);
+      const ok =
+        (installed && filters.status.has("installed")) ||
+        (!installed && filters.status.has("not-installed"));
+      if (!ok) return false;
+    }
+    if (filters.environment.size > 0) {
+      const env = envLabelByCatalog.get(item.id);
+      if (!env || !filters.environment.has(env)) return false;
+    }
+    if (
+      filters.author.size > 0 &&
+      (!item.authorName || !filters.author.has(item.authorName))
+    ) {
+      return false;
+    }
+    return true;
+  };
+
+  const sortItems = (list: CatalogItem[]) => {
+    switch (sort) {
+      case "name-desc":
+        return [...list].sort((a, b) => b.name.localeCompare(a.name));
+      case "newest":
+        return [...list].sort(
+          (a, b) =>
+            new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime(),
+        );
+      case "oldest":
+        return [...list].sort(
+          (a, b) =>
+            new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime(),
+        );
+      case "most-tools":
+        return [...list].sort(
+          (a, b) => (b.toolCount ?? 0) - (a.toolCount ?? 0),
+        );
+      default:
+        return [...list].sort((a, b) => a.name.localeCompare(b.name));
+    }
+  };
+
+  const allFilteredItems = sortItems(
     filterByLabels(
       filterCatalogItems(catalogItems || [], searchQueryFromUrl),
       parsedLabels,
-    ),
-  ).filter((item) => item.id !== ARCHESTRA_MCP_CATALOG_ID);
+    )
+      .filter((item) => item.id !== ARCHESTRA_MCP_CATALOG_ID)
+      .filter(matchesAdvancedFilters),
+  );
 
   const personalItems = allFilteredItems.filter(
     (item) => item.scope === "personal",
@@ -1500,10 +882,38 @@ export function InternalMCPCatalog({
           inputClassName="w-full bg-background/50 backdrop-blur-sm border-border/50 focus:border-primary/50 transition-colors pl-9"
         />
         <McpCatalogLabelFilter />
+        <RegistryFilterDropdown
+          label="Status"
+          options={STATUS_OPTIONS}
+          selected={filters.status}
+          onToggle={(value) => toggleFilter("status", value)}
+        />
+        {environmentOptions.length > 0 && (
+          <RegistryFilterDropdown
+            label="Environment"
+            options={environmentOptions}
+            selected={filters.environment}
+            onToggle={(value) => toggleFilter("environment", value)}
+          />
+        )}
+        {authorOptions.length > 0 && (
+          <RegistryFilterDropdown
+            label="Author"
+            options={authorOptions}
+            selected={filters.author}
+            onToggle={(value) => toggleFilter("author", value)}
+          />
+        )}
+        <RegistrySortMenu value={sort} onChange={setSort} />
       </div>
       {hasLabelFilters && (
         <LabelFilterBadges onRemoveLabel={handleRemoveLabel} />
       )}
+      <RegistryFilterChips
+        selected={filters}
+        onRemove={removeFilter}
+        onClearAll={clearAdvancedFilters}
+      />
       <div className="space-y-6">
         {personalItems.length > 0 && (
           <div className="space-y-3">
@@ -1515,7 +925,13 @@ export function InternalMCPCatalog({
                 const serverInfo = getInstalledServerInfo(item);
                 return (
                   <McpServerCard
-                    variant={cardVariantForServerType(item.serverType)}
+                    variant={
+                      item.serverType === "builtin"
+                        ? "builtin"
+                        : item.serverType === "remote"
+                          ? "remote"
+                          : "local"
+                    }
                     key={item.id}
                     item={item}
                     installedServer={serverInfo.installedServer}
@@ -1525,33 +941,16 @@ export function InternalMCPCatalog({
                       undefined
                     }
                     deploymentStatuses={deploymentStatuses}
-                    onInstallRemoteServer={() =>
-                      handleInstallRemoteServer(item, false)
-                    }
+                    onInstallRemoteServer={() => install.installRemote(item)}
                     onInstallLocalServer={() =>
                       isPlaywrightCatalogItem(item.id)
-                        ? handleInstallPlaywright(item)
-                        : handleInstallLocalServer(item)
+                        ? install.installPlaywright(item)
+                        : install.installLocal(item)
                     }
                     onReinstall={(flagged, options) =>
                       handleReinstall(item, flagged, options)
                     }
-                    onEdit={() => setEditingItem(item)}
-                    onDetails={() => {
-                      setDetailsServerName(item.name);
-                    }}
-                    onDelete={() => setDeletingItem(item)}
-                    onClone={() => handleClone(item)}
-                    onRestartPodsStarted={handleRestartPodsStarted}
-                    onRestartPodsFailed={handleRestartPodsFailed}
-                    onCancelInstallation={handleCancelInstallation}
-                    onAddPersonalConnection={() =>
-                      handleAddPersonalConnection(item)
-                    }
-                    onAddSharedConnection={(teamId) =>
-                      handleAddSharedConnection(item, teamId)
-                    }
-                    onAddOrgConnection={() => handleAddOrgConnection(item)}
+                    onCancelInstallation={install.cancelInstallation}
                     isBuiltInPlaywright={isPlaywrightCatalogItem(item.id)}
                   />
                 );
@@ -1572,7 +971,13 @@ export function InternalMCPCatalog({
                 const serverInfo = getInstalledServerInfo(item);
                 return (
                   <McpServerCard
-                    variant={cardVariantForServerType(item.serverType)}
+                    variant={
+                      item.serverType === "builtin"
+                        ? "builtin"
+                        : item.serverType === "remote"
+                          ? "remote"
+                          : "local"
+                    }
                     key={item.id}
                     item={item}
                     installedServer={serverInfo.installedServer}
@@ -1582,33 +987,16 @@ export function InternalMCPCatalog({
                       undefined
                     }
                     deploymentStatuses={deploymentStatuses}
-                    onInstallRemoteServer={() =>
-                      handleInstallRemoteServer(item, false)
-                    }
+                    onInstallRemoteServer={() => install.installRemote(item)}
                     onInstallLocalServer={() =>
                       isPlaywrightCatalogItem(item.id)
-                        ? handleInstallPlaywright(item)
-                        : handleInstallLocalServer(item)
+                        ? install.installPlaywright(item)
+                        : install.installLocal(item)
                     }
                     onReinstall={(flagged, options) =>
                       handleReinstall(item, flagged, options)
                     }
-                    onEdit={() => setEditingItem(item)}
-                    onDetails={() => {
-                      setDetailsServerName(item.name);
-                    }}
-                    onDelete={() => setDeletingItem(item)}
-                    onClone={() => handleClone(item)}
-                    onRestartPodsStarted={handleRestartPodsStarted}
-                    onRestartPodsFailed={handleRestartPodsFailed}
-                    onCancelInstallation={handleCancelInstallation}
-                    onAddPersonalConnection={() =>
-                      handleAddPersonalConnection(item)
-                    }
-                    onAddSharedConnection={(teamId) =>
-                      handleAddSharedConnection(item, teamId)
-                    }
-                    onAddOrgConnection={() => handleAddOrgConnection(item)}
+                    onCancelInstallation={install.cancelInstallation}
                     isBuiltInPlaywright={isPlaywrightCatalogItem(item.id)}
                   />
                 );
@@ -1641,79 +1029,16 @@ export function InternalMCPCatalog({
         )}
       </div>
 
-      <CreateCatalogDialog
-        isOpen={isDialogOpened("create")}
-        cloneValues={cloneValues ?? undefined}
-        clonedFrom={cloneSourceId ?? undefined}
-        onClose={() => {
-          setCloneValues(null);
-          setCloneSourceId(null);
-          closeDialog("create");
-        }}
-        onSuccess={(createdItem) => {
-          // Auto-open the appropriate install dialog based on server type
-          if (createdItem.serverType === "local") {
-            handleInstallLocalServer(createdItem);
-          } else if (createdItem.serverType === "remote") {
-            handleInstallRemoteServer(createdItem, false);
-          }
-          // For builtin servers, no connect dialog is needed
-        }}
-      />
-
       <CustomServerRequestDialog
         isOpen={isDialogOpened("custom-request")}
         onClose={() => closeDialog("custom-request")}
       />
 
-      <EditCatalogDialog
-        item={editingItem}
-        onClose={() => {
-          const item = editingItem;
+      {/* Shared install-mode dialogs (remote, OAuth, no-auth, local). */}
+      {install.dialogs}
 
-          if (item) {
-            setEditingItem(null);
-            const serverInfo = getInstalledServerInfo(item);
-            // Only auto-trigger reinstall if not already in error state
-            // (user should click "Reinstall Required" button to retry after error)
-            const isInErrorState =
-              serverInfo.installedServer?.localInstallationStatus === "error";
-            if (
-              serverInfo.installedServer?.reinstallRequired &&
-              !isInErrorState
-            ) {
-              // If the same edit also set catalogReinstallRequired (multi-tenant
-              // local catalog whose execution config changed), chain the catalog
-              // reinstall after the per-install one — otherwise the admin would
-              // see the catalog Reinstall button reappear and have to click it
-              // separately.
-              const alsoReinstallCatalog =
-                item.multitenant === true &&
-                item.catalogReinstallRequired === true;
-              handleReinstall(
-                item,
-                undefined,
-                alsoReinstallCatalog
-                  ? { alsoReinstallCatalog: true }
-                  : undefined,
-              );
-            }
-          }
-        }}
-      />
-
-      <DetailsDialog
-        onClose={() => {
-          setDetailsServerName(null);
-        }}
-        server={detailsServerData || null}
-      />
-
-      <DeleteCatalogDialog
-        item={deletingItem}
-        onClose={() => setDeletingItem(null)}
-      />
-
+      {/* Reinstall + reauth reuse the install dialog components but keep their
+          own instances/state so they stay independent of the install flow. */}
       <RemoteServerInstallDialog
         isOpen={isDialogOpened("remote-install")}
         onClose={() => {
@@ -1723,28 +1048,14 @@ export function InternalMCPCatalog({
           setReinstallServerId(null);
           setReinstallServerTeamId(null);
           setReinstallServerScope(undefined);
-          setPreselectedTeamId(null);
-          setInstallPersonalOnly(false);
-          setInstallOrgOnly(false);
         }}
-        onConfirm={handleRemoteServerInstallConfirm}
+        onConfirm={handleRemoteServerReauthOrReinstallConfirm}
         catalogItem={selectedCatalogItem}
-        isInstalling={
-          installMutation.isPending ||
-          reauthMutation.isPending ||
-          reinstallMutation.isPending
-        }
+        isInstalling={reauthMutation.isPending || reinstallMutation.isPending}
         isReauth={!!reauthServerId}
         isReinstall={!!reinstallServerId && !reauthServerId}
-        existingTeamId={
-          reauthServerId ? reauthExistingTeamId : reinstallServerTeamId
-        }
-        existingScope={
-          reauthServerId ? reauthExistingScope : reinstallServerScope
-        }
-        preselectedTeamId={preselectedTeamId}
-        personalOnly={installPersonalOnly}
-        orgOnly={installOrgOnly}
+        existingTeamId={reinstallServerTeamId}
+        existingScope={reinstallServerScope}
       />
 
       <OAuthConfirmationDialog
@@ -1755,20 +1066,13 @@ export function InternalMCPCatalog({
           }
         }}
         serverName={selectedCatalogItem?.name || ""}
-        onConfirm={handleOAuthConfirm}
+        onConfirm={handleReauthOAuthConfirm}
         onCancel={() => {
           closeDialog("oauth");
           setSelectedCatalogItem(null);
           setReauthServerId(null);
-          setPreselectedTeamId(null);
-          setInstallPersonalOnly(false);
-          setInstallOrgOnly(false);
         }}
         catalogId={selectedCatalogItem?.id}
-        preselectedTeamId={preselectedTeamId}
-        personalOnly={installPersonalOnly}
-        orgOnly={installOrgOnly}
-        isReauth={!!reauthServerId}
       />
 
       <ReinstallConfirmationDialog
@@ -1784,23 +1088,6 @@ export function InternalMCPCatalog({
         targets={reinstallFlaggedTargets}
       />
 
-      <NoAuthInstallDialog
-        isOpen={isDialogOpened("no-auth")}
-        onClose={() => {
-          closeDialog("no-auth");
-          setNoAuthCatalogItem(null);
-          setPreselectedTeamId(null);
-          setInstallPersonalOnly(false);
-          setInstallOrgOnly(false);
-        }}
-        onInstall={handleNoAuthConfirm}
-        catalogItem={noAuthCatalogItem}
-        isInstalling={installMutation.isPending}
-        preselectedTeamId={preselectedTeamId}
-        personalOnly={installPersonalOnly}
-        orgOnly={installOrgOnly}
-      />
-
       {localServerCatalogItem && (
         <LocalServerInstallDialog
           isOpen={isDialogOpened("local-install")}
@@ -1811,28 +1098,14 @@ export function InternalMCPCatalog({
             setReinstallServerTeamId(null);
             setReinstallServerScope(undefined);
             setReauthServerId(null);
-            setPreselectedTeamId(null);
-            setInstallPersonalOnly(false);
-            setInstallOrgOnly(false);
           }}
-          onConfirm={handleLocalServerInstallConfirm}
+          onConfirm={handleLocalServerReauthOrReinstallConfirm}
           catalogItem={localServerCatalogItem}
-          isInstalling={
-            installMutation.isPending ||
-            reinstallMutation.isPending ||
-            reauthMutation.isPending
-          }
+          isInstalling={reinstallMutation.isPending || reauthMutation.isPending}
           isReinstall={!!reinstallServerId}
-          existingTeamId={
-            reauthServerId ? reauthExistingTeamId : reinstallServerTeamId
-          }
-          existingScope={
-            reauthServerId ? reauthExistingScope : reinstallServerScope
-          }
+          existingTeamId={reinstallServerTeamId}
+          existingScope={reinstallServerScope}
           isReauth={!!reauthServerId}
-          preselectedTeamId={preselectedTeamId}
-          personalOnly={installPersonalOnly}
-          orgOnly={installOrgOnly}
         />
       )}
 
@@ -1846,21 +1119,21 @@ export function InternalMCPCatalog({
               (item) => item.id === manageCatalogId,
             );
             if (!catalogItem) return;
-            handleAddPersonalConnection(catalogItem);
+            install.addPersonalConnection(catalogItem);
           }}
           onAddSharedConnection={(teamId) => {
             const catalogItem = catalogItems?.find(
               (item) => item.id === manageCatalogId,
             );
             if (!catalogItem) return;
-            handleAddSharedConnection(catalogItem, teamId);
+            install.addSharedConnection(catalogItem, teamId);
           }}
           onAddOrgConnection={() => {
             const catalogItem = catalogItems?.find(
               (item) => item.id === manageCatalogId,
             );
             if (!catalogItem) return;
-            handleAddOrgConnection(catalogItem);
+            install.addOrgConnection(catalogItem);
           }}
         />
       )}
