+++ platform/frontend/src/app/mcp/registry/_parts/add-service-account-dialog.tsx
