+++ platform/frontend/src/app/mcp/registry/_parts/app-settings-dialog.tsx
