+++ platform/frontend/src/app/mcp/registry/_parts/catalog-setup-wizard.tsx
@@ -89,16 +89,16 @@ import {
   useToolsWithAssignments,
 } from "@/lib/tools/tool.query";
 import { cn } from "@/lib/utils";
-import { InstallationProgress } from "../../_parts/installation-progress";
+import { InstallationProgress } from "./installation-progress";
 import {
   LocalServerInstallDialog,
   type LocalServerInstallResult,
-} from "../../_parts/local-server-install-dialog";
+} from "./local-server-install-dialog";
+import type { CatalogItem } from "./mcp-server-card";
 import {
   RemoteServerInstallDialog,
   type RemoteServerInstallResult,
-} from "../../_parts/remote-server-install-dialog";
-import type { CatalogItem } from "./mcp-server-card";
+} from "./remote-server-install-dialog";
 
 export type SetupStepId = "configuration" | "test" | "tools";
 
@@ -334,7 +334,7 @@ export function TestConnectionStep({ item }: { item: CatalogItem }) {
             {needsRegistryHandoff ? (
               <Button asChild>
                 <Link
-                  href={`/mcp/registry/beta?${MCP_CATALOG_INSTALL_QUERY_PARAM}=${item.id}`}
+                  href={`/mcp/registry?${MCP_CATALOG_INSTALL_QUERY_PARAM}=${item.id}`}
                 >
                   Install
                 </Link>
@@ -373,9 +373,7 @@ export function TestConnectionStep({ item }: { item: CatalogItem }) {
               Retry
             </Button>
             <Button variant="ghost" size="sm" asChild>
-              <Link href={`/mcp/registry/beta/${item.id}?tab=logs`}>
-                View logs
-              </Link>
+              <Link href={`/mcp/registry/${item.id}?tab=logs`}>View logs</Link>
             </Button>
           </div>
         </div>
