+++ platform/frontend/src/app/mcp/registry/_parts/manage-users-dialog.tsx
@@ -6,28 +6,24 @@ import {
   E2eTestId,
   formatSecretStorageType,
   getDocsUrl,
-  getManageCredentialsAddToTeamOptionTestId,
   type McpDeploymentStatusEntry,
 } from "@archestra/shared";
 import { format } from "date-fns";
 import {
   AlertTriangle,
-  ChevronDown,
   Info,
   KeyRound,
-  PlugZap,
   Plus,
   RefreshCw,
   Trash,
   User,
   Zap,
 } from "lucide-react";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { toast } from "sonner";
 import { ExternalDocsLink } from "@/components/external-docs-link";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
 import {
   Dialog,
   DialogContent,
@@ -36,19 +32,6 @@ import {
   DialogStickyFooter,
   DialogTitle,
 } from "@/components/ui/dialog";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-  Empty,
-  EmptyContent,
-  EmptyDescription,
-  EmptyHeader,
-  EmptyMedia,
-} from "@/components/ui/empty";
 import {
   Popover,
   PopoverContent,
@@ -89,6 +72,7 @@ import {
 } from "@/lib/mcp/internal-mcp-catalog.query";
 import { useDeleteMcpServer, useMcpServers } from "@/lib/mcp/mcp-server.query";
 import { useMyTeams } from "@/lib/teams/team.query";
+import { AddServiceAccountDialog } from "./add-service-account-dialog";
 import { useCanModifyCatalogItem } from "./catalog-edit-access";
 import { type DeploymentState, DeploymentStatusDot } from "./deployment-status";
 import { formatOAuthFailureDetail } from "./oauth-reauth-detail";
@@ -155,7 +139,6 @@ interface ManageUsersContentProps {
   deploymentStatuses?: Record<string, McpDeploymentStatusEntry>;
   onOpenPodLogs?: (serverId: string) => void;
   hideHeader?: boolean;
-  bodyTestId?: string;
 }
 
 export function ManageUsersContent({
@@ -169,7 +152,6 @@ export function ManageUsersContent({
   deploymentStatuses = {},
   onOpenPodLogs,
   hideHeader = false,
-  bodyTestId,
 }: ManageUsersContentProps) {
   // Subscribe to live mcp-servers query to get fresh data. We fetch all
   // servers (no catalogId filter) and keep those installed from this catalog.
@@ -195,6 +177,9 @@ export function ManageUsersContent({
   const { data: hasMcpServerAdminPermission } = useHasPermissions({
     mcpServerInstallation: ["admin"],
   });
+
+  const [serviceAccountDialogOpen, setServiceAccountDialogOpen] =
+    useState(false);
 
   // Use the first server for display purposes
   const firstServer = allServers?.[0];
@@ -369,6 +354,51 @@ export function ManageUsersContent({
     return mcpServer.ownerEmail || "Deleted user";
   };
 
+  const split = splitByScope(allServers);
+  const canonicalStateByPod = computeCanonicalStateByPod(
+    allServers,
+    deploymentStatuses,
+  );
+
+  const personalRows: ConnectionRow[] = [
+    ...(split.myPersonalServer
+      ? [{ server: split.myPersonalServer, isYou: true } as const]
+      : []),
+    ...split.otherPersonalServers.map((s) => ({ server: s, isYou: false })),
+  ];
+  const serviceAccountRows: ConnectionRow[] = [
+    ...split.teamServers.map((s) => ({ server: s, isYou: false })),
+    ...split.orgServers.map((s) => ({ server: s, isYou: false })),
+  ];
+
+  const canAddPersonal =
+    hasAddCallbacks && !!onAddPersonalConnection && !split.myPersonalServer;
+  const canAddTeam =
+    hasAddCallbacks &&
+    !!onAddSharedConnection &&
+    split.availableTeamsForShared.length > 0;
+  const canAddOrg =
+    hasAddCallbacks &&
+    !!onAddOrgConnection &&
+    !split.hasOrgConnection &&
+    !!hasMcpServerAdminPermission;
+  const canAddServiceAccount = canAddTeam || canAddOrg;
+
+  const rowProps: RowRenderProps = {
+    isOAuthServer,
+    deploymentStatuses,
+    canonicalStateByPod,
+    getCredentialOwnerName,
+    canReauthenticate,
+    getReauthTooltip,
+    canRevoke,
+    getRevokeTooltip,
+    handleReauthenticate,
+    handleRevoke,
+    isDeleting: deleteMcpServerMutation.isPending,
+    onOpenPodLogs,
+  };
+
   return (
     <>
       {!hideHeader && (
@@ -384,101 +414,83 @@ export function ManageUsersContent({
         </DialogHeader>
       )}
 
-      <div
-        className={hideHeader ? "space-y-4 px-4 py-4" : "space-y-4 pb-4"}
-        data-testid={bodyTestId}
-      >
+      <div className={hideHeader ? "space-y-6" : "space-y-6 pb-4"}>
         {catalogItem && (
           <AgentConnectionsSection
             item={catalogItem}
             connections={allServers}
           />
         )}
-        {(() => {
-          const split = splitByScope(allServers);
-          const hasContent = allServers.length > 0;
 
-          const installMenu = (
-            <InstallMenuButton
-              onAddPersonal={
-                hasAddCallbacks &&
-                onAddPersonalConnection &&
-                !split.myPersonalServer
-                  ? () => {
-                      onClose();
-                      onAddPersonalConnection();
-                    }
-                  : undefined
-              }
-              onAddForTeam={
-                hasAddCallbacks && onAddSharedConnection
-                  ? (teamId) => {
-                      onClose();
-                      onAddSharedConnection(teamId);
-                    }
-                  : undefined
-              }
-              onAddForOrg={
-                hasAddCallbacks && onAddOrgConnection && !split.hasOrgConnection
-                  ? () => {
-                      onClose();
-                      onAddOrgConnection();
-                    }
-                  : undefined
-              }
-              availableTeamsForShared={split.availableTeamsForShared}
-              addOrgDisabled={!hasMcpServerAdminPermission}
-              addOrgDisabledReason={
-                !hasMcpServerAdminPermission
-                  ? "Only organization admins can install organization-wide"
-                  : undefined
-              }
-            />
-          );
+        {(personalRows.length > 0 || canAddPersonal) && (
+          <ConnectionsSection
+            title="Personal connections"
+            description="Each person connects their own account."
+            emptyText="No personal connections yet."
+            rows={personalRows}
+            tableTestId={E2eTestId.ManageCredentialsDialogTable}
+            action={
+              canAddPersonal ? (
+                <Button
+                  variant="outline"
+                  size="sm"
+                  className="h-7 text-xs"
+                  onClick={() => {
+                    onClose();
+                    onAddPersonalConnection?.();
+                  }}
+                >
+                  <Plus className="mr-1 h-3 w-3" />
+                  Connect my account
+                </Button>
+              ) : null
+            }
+            {...rowProps}
+          />
+        )}
 
-          return (
-            <div className="space-y-2">
-              <div className="flex items-center justify-between gap-4">
-                <div className="space-y-0.5">
-                  <h4 className="text-sm font-medium">Connections</h4>
-                  <p className="text-sm text-muted-foreground">
-                    Accounts connected to this server.
-                  </p>
-                </div>
-                {installMenu}
-              </div>
-              <Card>
-                <CardContent className="p-0">
-                  {hasContent ? (
-                    <UnifiedConnectionsTable
-                      myPersonalServer={split.myPersonalServer}
-                      otherPersonalServers={split.otherPersonalServers}
-                      teamServers={split.teamServers}
-                      orgServers={split.orgServers}
-                      isOAuthServer={isOAuthServer}
-                      getCredentialOwnerName={getCredentialOwnerName}
-                      canReauthenticate={canReauthenticate}
-                      getReauthTooltip={getReauthTooltip}
-                      canRevoke={canRevoke}
-                      getRevokeTooltip={getRevokeTooltip}
-                      handleReauthenticate={handleReauthenticate}
-                      handleRevoke={handleRevoke}
-                      isDeleting={deleteMcpServerMutation.isPending}
-                      deploymentStatuses={deploymentStatuses}
-                      onOpenPodLogs={onOpenPodLogs}
-                      availableTeamsForShared={split.availableTeamsForShared}
-                    />
-                  ) : (
-                    <p className="text-sm text-muted-foreground px-4 py-3">
-                      No callers yet.
-                    </p>
-                  )}
-                </CardContent>
-              </Card>
-            </div>
-          );
-        })()}
+        {(serviceAccountRows.length > 0 || canAddServiceAccount) && (
+          <ConnectionsSection
+            title="Service accounts"
+            description="Shared team & organization keys."
+            emptyText="No service accounts yet."
+            rows={serviceAccountRows}
+            tableTestId={E2eTestId.ManageServiceAccountsTable}
+            action={
+              canAddServiceAccount ? (
+                <Button
+                  variant="outline"
+                  size="sm"
+                  className="h-7 text-xs"
+                  data-testid={
+                    E2eTestId.ManageCredentialsAddServiceAccountButton
+                  }
+                  onClick={() => setServiceAccountDialogOpen(true)}
+                >
+                  <Plus className="mr-1 h-3 w-3" />
+                  Add service account
+                </Button>
+              ) : null
+            }
+            {...rowProps}
+          />
+        )}
       </div>
+
+      <AddServiceAccountDialog
+        open={serviceAccountDialogOpen}
+        onOpenChange={setServiceAccountDialogOpen}
+        availableTeams={canAddTeam ? split.availableTeamsForShared : []}
+        canAddOrg={canAddOrg}
+        onConfirm={(target) => {
+          onClose();
+          if (target.type === "org") {
+            onAddOrgConnection?.();
+          } else {
+            onAddSharedConnection?.(target.teamId);
+          }
+        }}
+      />
 
       {!hideHeader && (
         <DialogStickyFooter>
@@ -495,137 +507,12 @@ type ServerEntry = NonNullable<
   ReturnType<typeof useMcpServers>["data"]
 >[number];
 
-interface InstallMenuButtonProps {
-  onAddPersonal?: () => void;
-  onAddForTeam?: (teamId: string) => void;
-  onAddForOrg?: () => void;
-  availableTeamsForShared: Array<{ id: string; name: string }>;
-  addOrgDisabled?: boolean;
-  addOrgDisabledReason?: string;
-}
+type ConnectionRow = { server: ServerEntry; isYou: boolean };
 
-function InstallMenuButton({
-  onAddPersonal,
-  onAddForTeam,
-  onAddForOrg,
-  availableTeamsForShared,
-  addOrgDisabled,
-  addOrgDisabledReason,
-}: InstallMenuButtonProps) {
-  const teamItems =
-    onAddForTeam && availableTeamsForShared.length > 0
-      ? availableTeamsForShared.map((team) => ({
-          key: `team-${team.id}`,
-          label: `Install for ${team.name}`,
-          onClick: () => onAddForTeam(team.id),
-          testId: getManageCredentialsAddToTeamOptionTestId(team.name),
-        }))
-      : [];
-
-  const installItems = [
-    ...(onAddPersonal
-      ? [
-          {
-            key: "personal",
-            label: "Install for myself",
-            onClick: onAddPersonal,
-            testId: undefined as string | undefined,
-          },
-        ]
-      : []),
-    ...(onAddForOrg
-      ? [
-          {
-            key: "org",
-            label: "Install for organization",
-            onClick: onAddForOrg,
-            disabled: !!addOrgDisabled,
-            disabledReason: addOrgDisabledReason,
-            testId: E2eTestId.ManageCredentialsAddToOrgButton,
-          },
-        ]
-      : []),
-    ...teamItems,
-  ];
-
-  if (installItems.length === 0) return null;
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant="outline"
-          size="sm"
-          className="h-7 text-xs"
-          data-testid={E2eTestId.ManageCredentialsAddToTeamButton}
-        >
-          <Plus className="mr-1 h-3 w-3" />
-          Install
-          <ChevronDown className="ml-1 h-3 w-3" />
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end">
-        {installItems.map((item) => {
-          const disabled = "disabled" in item && item.disabled;
-          const reason =
-            "disabledReason" in item ? item.disabledReason : undefined;
-          const node = (
-            <DropdownMenuItem
-              key={item.key}
-              onClick={disabled ? undefined : item.onClick}
-              disabled={disabled}
-              data-testid={item.testId}
-            >
-              {item.label}
-            </DropdownMenuItem>
-          );
-          if (disabled && reason) {
-            return (
-              <TooltipProvider key={item.key}>
-                <Tooltip>
-                  <TooltipTrigger asChild>
-                    <div>{node}</div>
-                  </TooltipTrigger>
-                  <TooltipContent>{reason}</TooltipContent>
-                </Tooltip>
-              </TooltipProvider>
-            );
-          }
-          return <div key={item.key}>{node}</div>;
-        })}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-function UnifiedConnectionsTable({
-  myPersonalServer,
-  otherPersonalServers,
-  teamServers,
-  orgServers,
-  isOAuthServer,
-  getCredentialOwnerName,
-  canReauthenticate,
-  getReauthTooltip,
-  canRevoke,
-  getRevokeTooltip,
-  handleReauthenticate,
-  handleRevoke,
-  isDeleting,
-  deploymentStatuses = {},
-  onOpenPodLogs,
-  onAddPersonal,
-  availableTeamsForShared,
-  onAddForTeam,
-  onAddForOrg,
-  addOrgDisabled,
-  addOrgDisabledReason,
-}: {
-  myPersonalServer: ServerEntry | null;
-  otherPersonalServers: ServerEntry[];
-  teamServers: ServerEntry[];
-  orgServers: ServerEntry[];
+interface RowRenderProps {
   isOAuthServer: boolean;
+  deploymentStatuses: Record<string, McpDeploymentStatusEntry>;
+  canonicalStateByPod: Map<string, string>;
   getCredentialOwnerName: (s: ServerEntry) => string;
   canReauthenticate: (s: ServerEntry) => boolean;
   getReauthTooltip: (s: ServerEntry) => string;
@@ -634,165 +521,79 @@ function UnifiedConnectionsTable({
   handleReauthenticate: (s: ServerEntry) => void;
   handleRevoke: (s: ServerEntry) => void;
   isDeleting: boolean;
-  deploymentStatuses?: Record<string, McpDeploymentStatusEntry>;
   onOpenPodLogs?: (serverId: string) => void;
-  onAddPersonal?: () => void;
-  availableTeamsForShared: Array<{ id: string; name: string }>;
-  onAddForTeam?: (teamId: string) => void;
-  onAddForOrg?: () => void;
-  addOrgDisabled?: boolean;
-  addOrgDisabledReason?: string;
-}) {
-  const rows = [
-    ...(myPersonalServer
-      ? [{ server: myPersonalServer, isYou: true } as const]
-      : []),
-    ...otherPersonalServers.map((s) => ({ server: s, isYou: false }) as const),
-    ...teamServers.map((s) => ({ server: s, isYou: false }) as const),
-    ...orgServers.map((s) => ({ server: s, isYou: false }) as const),
-  ];
+}
 
+function ConnectionsSection({
+  title,
+  description,
+  action,
+  rows,
+  emptyText,
+  tableTestId,
+  ...rowProps
+}: {
+  title: string;
+  description: string;
+  action: React.ReactNode;
+  rows: ConnectionRow[];
+  emptyText: string;
+  tableTestId: string;
+} & RowRenderProps) {
+  return (
+    <div className="space-y-2">
+      <div className="flex items-start justify-between gap-4">
+        <div className="space-y-0.5">
+          <h4 className="text-sm font-medium">{title}</h4>
+          <p className="text-sm text-muted-foreground">{description}</p>
+        </div>
+        {action}
+      </div>
+      <div className="overflow-hidden rounded-lg border">
+        {rows.length > 0 ? (
+          <ConnectionsTable rows={rows} testId={tableTestId} {...rowProps} />
+        ) : (
+          <p className="text-sm text-muted-foreground px-4 py-3">{emptyText}</p>
+        )}
+      </div>
+    </div>
+  );
+}
+
+function ConnectionsTable({
+  rows,
+  testId,
+  isOAuthServer,
+  deploymentStatuses,
+  canonicalStateByPod,
+  getCredentialOwnerName,
+  canReauthenticate,
+  getReauthTooltip,
+  canRevoke,
+  getRevokeTooltip,
+  handleReauthenticate,
+  handleRevoke,
+  isDeleting,
+  onOpenPodLogs,
+}: {
+  rows: ConnectionRow[];
+  testId: string;
+} & RowRenderProps) {
   const hasDeploymentStatuses = rows.some(
     (r) => deploymentStatuses[r.server.id],
   );
 
-  // Multi-tenant catalogs alias one pod across N caller rows. Each row's
-  // K8sDeployment instance tracks its own state independently, so the row
-  // that didn't observe the pod first stays "pending" while the other goes
-  // "failed". Pick a canonical state per podName so all rows agree.
-  const STATE_PRIORITY: Record<string, number> = {
-    failed: 4,
-    running: 3,
-    succeeded: 3,
-    pending: 2,
-    not_created: 1,
-  };
-  const canonicalStateByPod = new Map<string, string>();
-  for (const { server } of rows) {
-    const entry = deploymentStatuses[server.id];
-    if (!entry?.podName) continue;
-    const current = canonicalStateByPod.get(entry.podName);
-    if (
-      !current ||
-      (STATE_PRIORITY[entry.state] ?? 0) > (STATE_PRIORITY[current] ?? 0)
-    ) {
-      canonicalStateByPod.set(entry.podName, entry.state);
-    }
-  }
-
-  const teamItems =
-    onAddForTeam && availableTeamsForShared.length > 0
-      ? availableTeamsForShared.map((team) => ({
-          key: `team-${team.id}`,
-          label: `Install for ${team.name}`,
-          onClick: () => onAddForTeam(team.id),
-          testId: getManageCredentialsAddToTeamOptionTestId(team.name),
-        }))
-      : [];
-
-  const installItems = [
-    ...(onAddPersonal
-      ? [
-          {
-            key: "personal",
-            label: "Install for myself",
-            onClick: onAddPersonal,
-            testId: undefined as string | undefined,
-          },
-        ]
-      : []),
-    ...(onAddForOrg
-      ? [
-          {
-            key: "org",
-            label: "Install for organization",
-            onClick: onAddForOrg,
-            disabled: !!addOrgDisabled,
-            disabledReason: addOrgDisabledReason,
-            testId: E2eTestId.ManageCredentialsAddToOrgButton,
-          },
-        ]
-      : []),
-    ...teamItems,
-  ];
-
-  const installMenu =
-    installItems.length === 0 ? null : (
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button
-            variant="outline"
-            size="sm"
-            className="h-7 text-xs"
-            data-testid={E2eTestId.ManageCredentialsAddToTeamButton}
-          >
-            <Plus className="mr-1 h-3 w-3" />
-            Install
-            <ChevronDown className="ml-1 h-3 w-3" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="end">
-          {installItems.map((item) => {
-            const disabled = "disabled" in item && item.disabled;
-            const reason =
-              "disabledReason" in item ? item.disabledReason : undefined;
-            const node = (
-              <DropdownMenuItem
-                key={item.key}
-                onClick={disabled ? undefined : item.onClick}
-                disabled={disabled}
-                data-testid={item.testId}
-              >
-                {item.label}
-              </DropdownMenuItem>
-            );
-            if (disabled && reason) {
-              return (
-                <TooltipProvider key={item.key}>
-                  <Tooltip>
-                    <TooltipTrigger asChild>
-                      <div>{node}</div>
-                    </TooltipTrigger>
-                    <TooltipContent>{reason}</TooltipContent>
-                  </Tooltip>
-                </TooltipProvider>
-              );
-            }
-            return <div key={item.key}>{node}</div>;
-          })}
-        </DropdownMenuContent>
-      </DropdownMenu>
-    );
-
-  if (rows.length === 0) {
-    return (
-      <Empty className="border rounded-md py-8">
-        <EmptyHeader>
-          <EmptyMedia variant="icon">
-            <PlugZap />
-          </EmptyMedia>
-          <EmptyDescription>No credentials yet.</EmptyDescription>
-        </EmptyHeader>
-        {installMenu && (
-          <EmptyContent className="flex-row justify-center">
-            {installMenu}
-          </EmptyContent>
-        )}
-      </Empty>
-    );
-  }
-
   return (
-    <Table data-testid={E2eTestId.ManageCredentialsDialogTable}>
+    <Table data-testid={testId}>
       <TableHeader>
         <TableRow>
-          <TableHead className="w-[220px]">Owner</TableHead>
+          <TableHead className="whitespace-nowrap">Owner</TableHead>
           {hasDeploymentStatuses && (
-            <TableHead className="w-[260px]">Pod</TableHead>
+            <TableHead className="whitespace-nowrap">Pod</TableHead>
           )}
-          <TableHead>Secret Storage</TableHead>
-          <TableHead>Created At</TableHead>
-          <TableHead>Action</TableHead>
+          <TableHead className="whitespace-nowrap">Secret Storage</TableHead>
+          <TableHead className="whitespace-nowrap">Created At</TableHead>
+          <TableHead className="whitespace-nowrap">Action</TableHead>
         </TableRow>
       </TableHeader>
       <TableBody>
@@ -810,9 +611,7 @@ function UnifiedConnectionsTable({
                       <TooltipTrigger>
                         <AlertTriangle className="h-4 w-4 text-amber-500 flex-shrink-0" />
                       </TooltipTrigger>
-                      <TooltipContent className="max-w-xs">
-                        <p className="font-medium">Needs re-authentication</p>
-                      </TooltipContent>
+                      <TooltipContent>Needs re-authentication</TooltipContent>
                     </Tooltip>
                   </TooltipProvider>
                 )}
@@ -881,8 +680,11 @@ function UnifiedConnectionsTable({
             <TableCell className="text-muted-foreground">
               {formatSecretStorageType(server.secretStorageType)}
             </TableCell>
-            <TableCell className="text-muted-foreground">
-              {format(new Date(server.createdAt), "PPp")}
+            <TableCell
+              className="whitespace-nowrap text-muted-foreground"
+              title={format(new Date(server.createdAt), "PPpp")}
+            >
+              {format(new Date(server.createdAt), "PP")}
             </TableCell>
             <TableCell>
               <div className="flex flex-col gap-1">
@@ -981,6 +783,39 @@ function UnifiedConnectionsTable({
       </TableBody>
     </Table>
   );
+}
+
+// Multi-tenant catalogs alias one pod across N caller rows. Each row's
+// K8sDeployment instance tracks its own state independently, so the row that
+// didn't observe the pod first stays "pending" while the other goes "failed".
+// Pick a canonical state per podName (across every connection for the catalog,
+// so the personal and service-account tables agree) so all rows match.
+const DEPLOYMENT_STATE_PRIORITY: Record<string, number> = {
+  failed: 4,
+  running: 3,
+  succeeded: 3,
+  pending: 2,
+  not_created: 1,
+};
+
+function computeCanonicalStateByPod(
+  servers: ServerEntry[],
+  deploymentStatuses: Record<string, McpDeploymentStatusEntry>,
+): Map<string, string> {
+  const canonicalStateByPod = new Map<string, string>();
+  for (const server of servers) {
+    const entry = deploymentStatuses[server.id];
+    if (!entry?.podName) continue;
+    const current = canonicalStateByPod.get(entry.podName);
+    if (
+      !current ||
+      (DEPLOYMENT_STATE_PRIORITY[entry.state] ?? 0) >
+        (DEPLOYMENT_STATE_PRIORITY[current] ?? 0)
+    ) {
+      canonicalStateByPod.set(entry.podName, entry.state);
+    }
+  }
+  return canonicalStateByPod;
 }
 
 // The catalog-level "agent connections" setting as a standard settings row:
