+++ platform/frontend/src/app/mcp/registry/_parts/registry-list-controls.tsx
