+++ platform/frontend/src/app/mcp/registry/_parts/use-chat-with-catalog-item.ts
