--- platform/frontend/src/app/mcp/registry/beta/page.client.tsx
@@ -1,22 +1,0 @@
-"use client";
-
-import type { archestraApiTypes } from "@archestra/shared";
-import { InternalMCPCatalog } from "./_parts/InternalMCPCatalog";
-
-export default function McpRegistryBetaClient({
-  initialData,
-}: {
-  initialData: {
-    catalog: archestraApiTypes.GetInternalMcpCatalogResponses["200"];
-    servers: archestraApiTypes.GetMcpServersResponses["200"];
-  };
-}) {
-  return (
-    <div>
-      <InternalMCPCatalog
-        initialData={initialData.catalog}
-        installedServers={initialData.servers}
-      />
-    </div>
-  );
-}
