+++ platform/frontend/src/app/mcp/registry/catalog/[name]/page.tsx
