+++ platform/frontend/src/app/mcp/registry/catalog/[name]/readme-markdown.tsx
